types.ts作成
