+++ frontend/features/TaskDetail/Presenter.tsx
@@ -1,38 +1,14 @@
 "use client";
 
 import { colors } from "@/app/theme/colors";
-import { Status, statusLabel } from "@/constants/tasks";
+import { statusLabel } from "@/constants/tasks";
 import { Box, Button, Card, CardContent, Typography } from "@mui/material";
 import Link from "next/link";
+import { Task } from "./types";
 
 type Props = {
   task: Task;
   goalId?: number;
-};
-
-type Task = {
-  id: number;
-  goal_id: number;
-  title: string;
-  content: string;
-  due_date: string;
-  priority: string;
-  status: Status;
-  completed_at: string;
-  units?: Unit[];
-};
-
-export type Unit = {
-  id: number;
-  course_id: number;
-  unit_name: string;
-  course: Course;
-};
-
-export type Course = {
-  id: number;
-  level_number: number;
-  level_name: string;
 };
 
 export const Presenter = ({ task, goalId }: Props) => {
